+++ src/components/layout/list/Item.tsx
@@ -4,10 +4,14 @@ import classnames from "classnames";
 import styles from "./Item.module.scss";
 import { Value } from "./Value.js";
 import { Label } from "./Label.js";
+import { makeVariants } from "../../../utils/makeVariants";
+import { ICON, Icon } from "../../icons/Icon";
 
 interface LinkProps {
     href: string;
 }
+
+type Variant = "inset";
 
 interface Props {
     href?: string;
@@ -16,6 +20,8 @@ interface Props {
     className?: string;
     Link?: React.ComponentClass<LinkProps> | React.FC<LinkProps>;
     ratio?: string; // @TODO number/number/number/.. type?
+    variant?: Variant | Variant[];
+    selected?: boolean | undefined;
 }
 
 interface SubComponents {
@@ -23,10 +29,25 @@ interface SubComponents {
     Value: typeof Value;
 }
 
-// eslint-disable-next-line max-statements
+const icon = <Icon className={styles.icon} name={ICON.forward} />;
+const noIcon = <span className={styles.icon} />;
+
+// eslint-disable-next-line max-statements,max-lines-per-function
 const Item: React.FC<Props> & SubComponents = (props) => {
-    const { className, children, href, to, onClick, Link, ratio, ...restProps } = props;
+    const { className, children, href, to, onClick, Link, ratio, variant, selected, ...restProps } = props;
+    const v = makeVariants(variant);
     const cls = classnames(styles.item, className);
+
+    const innerCls = classnames(styles.content, {
+        [styles.inset]: v.includes("inset"),
+        [styles.selected]: selected,
+    });
+
+    const pre = typeof selected === "boolean"
+        ? (
+            selected ? icon : noIcon
+        )
+        : null;
 
     const liProps = {
         className: cls,
@@ -42,7 +63,7 @@ const Item: React.FC<Props> & SubComponents = (props) => {
         if (index in r) {
             flex = r[index];
         }
-        return <div style={{ flex }}>{child}</div>;
+        return <div style={{ flex }}>{pre}{child}</div>;
     });
 
     if (to) {
@@ -50,25 +71,25 @@ const Item: React.FC<Props> & SubComponents = (props) => {
             throw new TypeError("`to` prop given without `Link` component");
         }
 
-        return <li {...liProps}><Link href={to} {...restProps}><a className={styles.content}>{ren}</a></Link></li>;
+        return <li {...liProps}><Link href={to} {...restProps}><a className={innerCls}>{ren}</a></Link></li>;
     }
     if (href) {
         const aProps: typeof restProps & { onClick?: Props["onClick"]} = { ...restProps };
         if (onClick) {
             aProps.onClick = onClick;
         }
-        return <li {...liProps}><a href={href} {...aProps} className={styles.content}>{ren}</a></li>;
+        return <li {...liProps}><a href={href} {...aProps} className={innerCls}>{ren}</a></li>;
     }
     if (onClick) {
         return (
             <li {...liProps}>
-                <button className={classnames(styles.button, styles.content)} onClick={props.onClick}{...restProps}>
+                <button className={classnames(styles.button, innerCls)} onClick={props.onClick} {...restProps}>
                     {ren}
                 </button>
             </li>
         );
     }
-    return <li {...liProps}><div {...restProps} className={styles.content}>{ren}</div></li>;
+    return <li {...liProps}><div {...restProps} className={innerCls}>{ren}</div></li>;
 };
 
 Item.Label = Label;