+++ src/components/ui/toaster/types.ts
@@ -1,0 +1,9 @@
+interface Toast {
+    id: string;
+    text: string;
+    hide: boolean;
+}
+
+export type {
+    Toast,
+};