+++ src/demo/Demo.tsx
@@ -4,13 +4,16 @@ import styles from "./Demo.module.scss";
 import { Menu } from "./Menu";
 import { Main } from "./Main";
 import { componentsMap } from "./componentsMap";
+import { ToasterProvider } from "../components/ui/toaster/Toaster";
 
 const Demo: React.FC = (props) => {
     return (
-        <div className={styles.container}>
-            <Menu list={componentsMap} />
-            <Main />
-        </div>
+        <ToasterProvider>
+            <div className={styles.container}>
+                <Menu list={componentsMap} />
+                <Main />
+            </div>
+        </ToasterProvider>
     );
 };