+++ src/index.ts
@@ -19,3 +19,6 @@ export * from "./components/ui/action/Action.js";
 export * from "./components/ui/action/EqualActions.js";
 export * from "./components/ui/button/Button.js";
 export * from "./components/ui/directionPad/Pad.js";
+export * from "./components/ui/modal/Modal.js";
+export * from "./components/ui/modal/ModalButtons.js";
+export * from "./components/ui/toaster/Toaster.js";